Read the file to be uploaded
Begin the upload script by testing to make sure, we get the filename
and some data back from fs.readFile.

Use an empty string as a default argument for the filepath, so that if
we don't pass an argument when running the script, we get a sensible
error message (file not found).